requirements.txt
openai==1.13.3
SpeechRecognition==3.10.0
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
pyaudio==0.2.14
